+++ core/language/nl/gallery_acp.php
@@ -319,6 +319,8 @@ $lang = array_merge($lang, [
 	'RECENT_ON_INDEX'          => 'Laat recente afbeeldingen zien',
 	'RECENT_ON_INDEX_EXP'      => 'Wil je recente afbeeldingen op de galerij index weergeven',
 	'RECENT_ON_INDEX_COUNT'    => 'Recente afbeeldingsteller',
+	'FORUM_INDEX_IMAGES'       => 'Galerijafbeeldingen op de forumindex',
+
 	'PHPBB_INTEGRATION'        => 'phpBB integratie',
 	'PNG_ALLOWED'              => 'Bevoegt om PNG bestanden te uploaden',
 	'WEBP_ALLOWED'             => 'Bevoegt om WEBP bestanden te uploaden',