+++ core/language/ru/gallery_acp.php
@@ -304,6 +304,8 @@ $lang = array_merge($lang, [
 	'RECENT_ON_INDEX'                      => 'Включить последние фото',
 	'RECENT_ON_INDEX_EXP'                  => 'Отображение последних фотографий на главной странице галереи',
 	'RECENT_ON_INDEX_COUNT'                => 'Количество последних фото',
+	'FORUM_INDEX_IMAGES'                   => 'Изображения галереи на главной странице форума',
+
 	'PHPBB_INTEGRATION'                    => 'Интеграция в phpBB',
 	'PNG_ALLOWED'                          => 'PNG разрешён',
 	'WEBP_ALLOWED'                         => 'WEBP разрешён',